Server: Add batch concept to importing
A batch represents a logical unit of data that was fetched together.
This way, we can always return the most recent data from /stocks instead
of having to filter database entries by time. We can
also avoid duplicate entries returned from /stocks which previously
happened if data was fetched multiple times in a 24-hour period.

Also, check if the data most recent data is from an old batch on
startup. This means data is always going to be fetched after downtime
instead of having to do it manually or waiting for the fetching cronjob
to run.