Client: Make i18n uses type safe
Now all CountryCodes and ColumnIds must have translations to compile and
all uses of the "t" translation hook are checked to be valid.

This revealed an issue with epsGrowthThreeYearAverage not having a
translation. It turns out this field is not rendered anyway so removing
it from ColumnIds instead.

Also split up footer text into two different translated sections to
avoid using the "dangerouslySetInnerHTML" prop.