CI: Bump actions' versions
Currently, when running CI, 23 warnings are logged about various
deprecations. This should fix most of them.

Also, deploying to the server did not work because of issues with ssh:
  drone-scp error:  error copy file to dest: ***, error message: ssh: handshake failed: ssh: unable to authenticate, attempted methods [none publickey], no supported methods remain

I upgraded the server to Ubuntu 22.04, which comes with a newer
version of OpenSSH. This version does not play nicely with older
versions of Go, which scp-action@v0.1.2 is based on so bump the
version to 0.1.3.