Server: Enable IPv6 networking in production docker-compose.yml
The production server is running IPv6-only and has been broken for a
while because by default, Docker refuses to do IPv6 networking. This can
be fixed by adding a custom network to the compose file. Java will still
try to use IPv6, but this can be worked around by setting the
_JAVA_OPTIONS environment variable.