Client: Migrate to Vite
create-react-app is deprecated and the official suggestion is to move to
either a framework or a build tool like Vite. Since this is a basic SPA,
there's no need for a framework and Vite is a popular build tool that's
nice to use so I picked that option.

The migration was made easier by using the viject tool.